Rework declaration and exporting some more.
There's no need to pre-declare names. Each name only has one opportunity to be
typoed, and that's the call to declare().

After that, we use t() to get the type, which will catch typos.

Distinguish between exportable and internal-only types for a type
library. When TypeLib::A imports the Builtin types, it should not
automatically re-export them.