Fix a nasty heisenbug with the Sub::Quoted sub for coercions
We need to make sure that the generated code always ends with a single
statement containing the coerced value. With Moo, the generated code is
inlined into a do block, so if the last value of the block _isn't_ the coerced
value, then the block ends up returning an undef.