Mark TODO.pod as "Not for installation"
Due to historical reasons, EUMM installs all files called "*.pm"
or "*.pod" in the distributions root directory to the the install
target.

For instance,

    Dist-Foo-Bar/Bar.pm   -> ${site_lib}/Dist/Foo/Bar.pm
    Dist-Foo-Bar/TODO.pod -> ${site_lib}/Dist/Foo/TODO.pod
    Dist-Foo/TODO.pod     -> ${site_lib}/Dist/TODO.pod
    Dist/TODO.pod         -> ${site_lib}/TODO.pod

INSTALL.SKIP is used by ExtUtils::Install to mark things as "not for
installation" while performing "make install", and acts to filter
things in "blib/" from landing in ${site_lib}

It will still *appear* to have no impact because EUMM will still spit
out that its copying it from Dist-Foo/ to Dist-Foo/blib/

But it shouldn't end up in the install target.

Additionally, this rule filters out any resulting man-pages as well,
because TODO.pod will be manified before ExtUtils::Install gets a
chance.

And the proof is in the pudding.

  cpanm -l /tmp/elib Specio-$VERSION.tar.gz
  find /tmp/elib | grep -i TODO

Now should emit nothing.