Only fire one coercion in inlined coerce
Using an if condition for each coercion meant that all of the coercions could
fire when coercing a value. It also is wasteful, as the type checks will be
performed for all coerce types rather than stopping once the coercion has been
done.

Switch to using ternaries rather than if conditions.