Fix by is_class_loaded function
is_class_loaded in Types::Standard has a flaw in it (see
<https://rt.cpan.org/Ticket/Display.html?id=123408>), because of
which it does not work with the CV-in-stash optimisation.  This module
borrows the code and has the same flaw, resulting in a ‘Not a GLOB
reference’ error.  This patch (practically the same as the one in
Type-Tiny’s RT queue) fixes the problem.

For more information on the optimisation, which
will likely be included in perl 5.28, see:
https://rt.perl.org/Ticket/Display.html?id=132252#txn-1500037